Andrew Parise - extracted out the simulation test usage of the data structure to another package
