+++ queue/.gitignore
@@ -1,0 +1,41 @@
+# Created by .ignore support plugin (hsz.mobi)
+### Java template
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+# ignore intellij autogenerated
+*.iml
+
+# ignore flattened poms
+.flattened-pom.xml
+
+# ignore maven build output
+target/
+
+# ignore pom backup
+pom.xml.versionsBackup
+
+# ignore maven release files
+pom.xml.releaseBackup
+release.properties
+
