+++ queue/pom.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>queue</artifactId>
+    <version>1.0.3-SNAPSHOT</version>
+    <name>Queue Data Structure</name>
+    <description>Queue Data Structure Implementation</description>
+    <packaging>jar</packaging>
+
+    <parent>
+        <groupId>com.parisesoftware</groupId>
+        <artifactId>queue-parent</artifactId>
+        <version>1.0.3-SNAPSHOT</version>
+    </parent>
+
+    <properties>
+        <parisesoftware-linkedlist.version>1.0.3</parisesoftware-linkedlist.version>
+    </properties>
+
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.parisesoftware</groupId>
+            <artifactId>linkedlist</artifactId>
+            <version>${parisesoftware-linkedlist.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations-java5</artifactId>
+            <version>RELEASE</version>
+            <scope>compile</scope>
+        </dependency>
+
+    </dependencies>
+
+</project>
